arm64: dts: qcom: sdm630: add clocks for digital and analog audio codecs
The digital and analog codecs communicate with each other, synchronized
by an MCLK, which cannot turn off before the codecs power down. The bit
clock is better managed by the sound card driver instead of the digital
codec, since it (to my knowledge) does not power down the clock. Add the
Q6AFE clock controller to connect the MCLK with the codecs, and a fixed
clock to model a fake bit clock for consumption by the digital codec.

Signed-off-by: Richard Acayan <mailingradian@gmail.com>